numfmt: Add support for --delimiter and --field
Adds support for --delimiter field with support for default delimiter
parsing which is useful for piping in the output of utilities like ls
and df into numfmt.
--field argument support cut style arguments N-M which allows user to
specify which columns to apply conversion logic to. Useful for parsing
lines which contain a mix of numeric and non-numeric fields (like ls
-al)