maint/CICD ~ speedup CICD by decreasing redundant Travis CI workload
- 10 minutes to CI success/failure response from Travis CI (approx median time); down from 60 minutes

.# [why]

Most of the testing done on Travis CI is redundant with that done in the GHA:CICD.
With prior settings, Travis CI was taking about an hour (median) to report success/failure.
In contrast, the GHA:CICD workflow usually completes in about 13 minutes (median).

Travis CI job times are highly variable taking between 3 and 20 minutes per build.
Builds are run with some parallelism, and a report is submitted as soon as the
non-"allow failure" jobs are completed.