wc: print counts for each file as soon as computed
Change the behavior of `wc` to print the counts for a file as soon as
it is computed, instead of waiting to compute the counts for all files
before writing any output to `stdout`. The new behavior matches the
behavior of GNU `wc`.

The old behavior looked like this (the word "hello" is entered on
`stdin`):

    $ ./target/debug/coreutils wc tests/fixtures/wc/emptyfile.txt -
    hello
     0 0 0 tests/fixtures/wc/emptyfile.txt
     1 1 6
     1 1 6 total

The new behavior looks like this:

    $ ./target/debug/coreutils wc tests/fixtures/wc/emptyfile.txt -
     0 0 0 tests/fixtures/wc/emptyfile.txt
    hello
     1 1 6
     1 1 6 total