tac: read files in reverse in chunks
Change the implementation of `tac` when reading from a file: read the
file in chunks, starting from the end of the file, and stitching
together lines as they are discovered.

This change also fixes a bug in the behavior of `tac -b`. Previously, it
was not assuming that the line separators were at the *beginning* of
each line, which caused its output to disagree with GNU `tac`.

Fixes #2262.