chown: fix argument parsing when using `--reference`
chown can be used either with `--reference=RFILE FILE...` or with
`OWNER[:GROUP] FILE...`. In the former case, using `--reference` with a
single FILE argument raises an error as OWNER and FILE are both required
arguments.

Similarly to what has been done in chmod, setting FILE to be a required
argument only when `--reference` isn't used, and adding the contents of
OWNER to the file list solves this issue.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>