sort: fix `--key` argument handling
When using `sort` with the `--key` option, all subsequent arguments
are considered its values, leading to failures such as:

  $ sort -t ":" -k 3 /etc/passwd
  sort: failed to parse key `/etc/passwd`: invalid option: `e`

This is a side-effect of having `multiple(true)` for this option,
as `multiple` supports separating values with either a comma or
whitespace. Therefore, as long as no other flag or option is
provided, all arguments following `-k` are treated like values
for this option.

Switching to `use_delimiter(true)` instead ensures only
comma-separated values are taken into account, and the following
arguments are properly interpreted.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>