completions: completions by subcommand
Adds a (hidden) subcommand to the coreutils binary that generates
completions for a utility and shell to stdout.
use: coreutils completion <utility> <shell>

All utilities are now required to have a public `app` module with a
`get_app` function (the exact convention could still be changed,
there just needs to be a function that returns an `App` for each
utility in a place we can find).