refactor: make executable!() return the real binary name
executable!() no longer returns the binary name but the name
of the executable that was called.

The busybox style executable is a special case because when we're
called like `coreutils <utility>` the executable name is set to the
utility, not "coreutils". In all other cases the executable name is
the first command line argument.

To get the name of the utility regardless of the executable name
a new macro `utility_name!()` was added. It returns the crate name
minus the leading "uu_".

require manually setting the executable