error: Derive `Debug` for `UError`
Derives the `Debug` trait for the `UError` type. This satisfies the trait
bounds required by `std::result::Result::unwrap` and allows code to call
`unwrap()` on `UResult` and the enclosing `UError`.

While it isn't especially recommended to use `unwrap()` in code in general, it
proves to be very useful in tests and doctests, where explicit error checking
isn't needed as a panic will make the test fail, which is to be expected when
`unwrap()` is called.