realpath: add -L and -P command-line arguments
Add the `-L` (logical) and `-P` (physical) command-line arguments to
`realpath`. The `-L` option makes the program resolve ".." components
before symbolic links. The `-P` option makes the program resolve
symbolic links as they are encountered.

This also changes the default behavior of `realpath` from the `-L`
behavior to the `-P` behavior. This matches the behavior of GNU
`realpath`.