seq: use bigdecimal for floating point precision
Fix a floating-point precision issue with the update in the main loop of
`seq`. Before this commit, the behavior was

    $ seq 0.1 -0.1 -0.2
    0.1
    0.0
    -0.1

which was incorrect: -0.2 should be printed as well. This was happening
because the update in the main loop was essentially

    value = start + (i * increment);

which caused `value` to become `-0.20000000000000004`. That exceeds the
lower bound of -0.2, so that number was not printed. By using the
`bigdecimal` crate to represent floats of (practically) arbitrary
precision, we eliminate the issue.